+++ docs/en/changes/changes.md
@@ -5,6 +5,7 @@
 
 #### OAP Server
 * Fix wrong indices in the eBPF Profiling related models.
+* Support exclude the specific namespaces traffic in the eBPF Access Log receiver.
 
 
 #### UI